% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_generics.R
\name{augment}
\alias{augment}
\title{Augment a Fitted Object}
\usage{
augment(x, ...)
}
\arguments{
\item{x}{A fitted object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Typically a tibble with augmented results.
}
\description{
Lightweight alternative to broom's \code{augment()} tailored for this package.
}
