% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_structure_spec.R
\name{augment.term_structure_fit}
\alias{augment.term_structure_fit}
\title{Augment Term Structure Calibration}
\usage{
\method{augment}{term_structure_fit}(x, ...)
}
\arguments{
\item{x}{A \code{term_structure_fit} produced by \code{\link[=fit]{fit()}}.}

\item{...}{Unused.}
}
\value{
Tibble containing the calibrated term structure.
}
\description{
Returns the calibrated discount factors, zero rates, and forward rates in a
tidy tibble.
}
