% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bshw_model.R
\name{bshw_equivalent_volatility}
\alias{bshw_equivalent_volatility}
\title{Equivalent BSHW Volatility}
\usage{
bshw_equivalent_volatility(
  maturity,
  equity_vol,
  short_rate_vol,
  correlation,
  mean_reversion,
  integration_points = 2000L
)
}
\arguments{
\item{maturity}{Numeric maturity.}

\item{equity_vol}{Equity volatility parameter.}

\item{short_rate_vol}{Hull-White short-rate volatility.}

\item{correlation}{Equity/short-rate correlation.}

\item{mean_reversion}{Hull-White mean reversion.}

\item{integration_points}{Number of trapezoidal steps.}
}
\value{
Numeric scalar volatility.
}
\description{
Integrate the instantaneous forward volatility curve implied by the hybrid
model to obtain the Black-76 equivalent volatility.
}
