% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompFinanceR-package.R
\docType{package}
\name{CompFinanceR}
\alias{CompFinanceR-package}
\alias{CompFinanceR}
\title{CompFinanceR package documentation}
\description{
Core documentation stub generated to ensure compiled code exports are registered
and to provide a home for package-level metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Samcil/Computational-Finance-Course}
  \item Report bugs at \url{https://github.com/Samcil/Computational-Finance-Course/issues}
}

}
\author{
\strong{Maintainer}: Repository Contributor \email{contributor@example.com}

Authors:
\itemize{
  \item Lech Grzelak \email{lech.grzelak@example.com}
}

}
\keyword{internal}
