% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_generics.R
\name{fit}
\alias{fit}
\title{Fit a Model Specification}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{Model specification object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The result of the dispatched method.
}
\description{
Generic S3 function mirroring tidymodels semantics. Specifications created
within this package (e.g., \code{\link[=term_structure_spec]{term_structure_spec()}}) provide dedicated
methods.
}
\examples{
\dontrun{
fit(gbm_spec(initial_value = 100, drift = 0.05, volatility = 0.2), data)
}

}
