% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_models.R
\name{fx_h1_hw_characteristic_function}
\alias{fx_h1_hw_characteristic_function}
\title{FX Heston-Hull-White Characteristic Function}
\usage{
fx_h1_hw_characteristic_function(
  u,
  maturity,
  state,
  integration_points = 1500L
)
}
\arguments{
\item{u}{Numeric integration vector.}

\item{maturity}{Time to maturity.}

\item{state}{Internal state returned by \code{\link[=fx_h1_hw_spec]{fx_h1_hw_spec()}}.}

\item{integration_points}{Integer number of trapezoidal steps.}
}
\value{
Complex vector of characteristic function evaluations.
}
\description{
FX Heston-Hull-White Characteristic Function
}
