% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_models.R
\name{fx_h1_hw_spec}
\alias{fx_h1_hw_spec}
\title{FX Heston-Hull-White Specification}
\usage{
fx_h1_hw_spec(
  spot_fx,
  domestic_short_rate,
  foreign_short_rate,
  initial_variance,
  variance_mean_reversion,
  variance_long_term,
  variance_volatility,
  correlation_eq_var,
  correlation_eq_domestic,
  correlation_eq_foreign,
  correlation_var_domestic,
  correlation_var_foreign,
  correlation_domestic_foreign
)
}
\arguments{
\item{spot_fx}{Spot FX rate (domestic per foreign unit).}

\item{domestic_short_rate}{Hull-White domestic short-rate spec.}

\item{foreign_short_rate}{Hull-White foreign short-rate spec.}

\item{initial_variance}{Initial Heston variance.}

\item{variance_mean_reversion}{Heston mean reversion (kappa).}

\item{variance_long_term}{Long-term variance level (theta).}

\item{variance_volatility}{Volatility of variance (gamma).}

\item{correlation_eq_var}{Correlation between equity and variance drivers.}

\item{correlation_eq_domestic}{Correlation between equity and domestic rate drivers.}

\item{correlation_eq_foreign}{Correlation between equity and foreign rate drivers.}

\item{correlation_var_domestic}{Correlation between variance and domestic rate drivers.}

\item{correlation_var_foreign}{Correlation between variance and foreign rate drivers.}

\item{correlation_domestic_foreign}{Correlation between domestic and foreign rate drivers.}
}
\value{
A \code{fx_h1_hw_spec} inheriting from \code{hybrid_spec}.
}
\description{
FX Heston-Hull-White Specification
}
