% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_paths_engines.R
\name{generate_standardized_normals}
\alias{generate_standardized_normals}
\title{Generate Standardized Normal Random Variables}
\usage{
generate_standardized_normals(n_paths, n_steps)
}
\arguments{
\item{n_paths}{Integer. Number of paths.}

\item{n_steps}{Integer. Number of time steps.}
}
\value{
Matrix of size (n_paths x n_steps) with standardized normal samples.
}
\description{
Generates a matrix of random normal samples and standardizes each time step
to have mean 0 and variance 1.
}
\keyword{internal}
