% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_models.R
\name{h1_hw_characteristic_function}
\alias{h1_hw_characteristic_function}
\title{Heston-Hull-White Characteristic Function}
\usage{
h1_hw_characteristic_function(
  u,
  maturity,
  theta_fun,
  initial_rate,
  mean_reversion,
  short_rate_vol,
  initial_variance,
  variance_mean_reversion,
  variance_long_term,
  variance_volatility,
  correlation_eq_var,
  correlation_eq_rate,
  integration_points = 2000L
)
}
\arguments{
\item{u}{Numeric integration vector.}

\item{maturity}{Time to maturity.}

\item{theta_fun}{Hull-White theta function.}

\item{initial_rate}{Initial short rate.}

\item{mean_reversion}{Hull-White mean reversion.}

\item{short_rate_vol}{Hull-White volatility (eta).}

\item{initial_variance}{Heston initial variance.}

\item{variance_mean_reversion}{Heston kappa.}

\item{variance_long_term}{Heston theta.}

\item{variance_volatility}{Heston gamma.}

\item{correlation_eq_var}{rho_{xv}.}

\item{correlation_eq_rate}{rho_{xr}.}

\item{integration_points}{Trapezoid steps for deterministic integrals.}
}
\value{
Complex vector of characteristic function values.
}
\description{
Mirrors \code{ChFH1HWModel} from the Python lectures.
}
