% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_models.R
\name{h1_hw_spec}
\alias{h1_hw_spec}
\title{Heston-Hull-White Hybrid Specification}
\usage{
h1_hw_spec(
  spot,
  short_rate,
  initial_variance,
  variance_mean_reversion,
  variance_long_term,
  variance_volatility,
  correlation_eq_var,
  correlation_eq_rate
)
}
\arguments{
\item{spot}{Numeric. Current equity spot price.}

\item{short_rate}{A \code{short_rate_spec} constructed with \code{model = "hull_white"}.}

\item{initial_variance}{Initial Heston variance.}

\item{variance_mean_reversion}{Heston mean reversion (kappa).}

\item{variance_long_term}{Long-term variance level (theta).}

\item{variance_volatility}{Volatility of variance (gamma).}

\item{correlation_eq_var}{Correlation between equity and variance drivers.}

\item{correlation_eq_rate}{Correlation between equity and short-rate drivers.}
}
\value{
A \code{h1_hw_spec} inheriting from \code{hybrid_spec}.
}
\description{
Heston-Hull-White Hybrid Specification
}
