% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortgage_products.R
\name{mortgage_schedule}
\alias{mortgage_schedule}
\title{Extract Mortgage Schedule}
\usage{
mortgage_schedule(object)
}
\arguments{
\item{object}{A \code{mortgage_annuity_spec} produced by
\code{\link[=mortgage_annuity_spec]{mortgage_annuity_spec()}}.}
}
\value{
Tibble containing the mortgage schedule.
}
\description{
Return the amortisation schedule stored on a mortgage specification.
}
