% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_utils.R
\name{new_diffusion_spec}
\alias{new_diffusion_spec}
\title{Create a Diffusion Process Specification}
\usage{
new_diffusion_spec(class, args, process_type, inheritance = character())
}
\description{
Convenience wrapper above \code{new_process_spec()} for continuous diffusions.
}
\keyword{internal}
