% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_utils.R
\name{new_jump_diffusion_spec}
\alias{new_jump_diffusion_spec}
\title{Create a Jump-Diffusion Specification}
\usage{
new_jump_diffusion_spec(class, args, process_type, inheritance = character())
}
\description{
Ensures jump diffusions inherit diffusion behaviour while recording jump
metadata.
}
\keyword{internal}
