% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_utils.R
\name{new_jump_spec}
\alias{new_jump_spec}
\title{Create a Jump Process Specification}
\usage{
new_jump_spec(class, args, process_type, inheritance = character())
}
\description{
Create a Jump Process Specification
}
\keyword{internal}
