% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_utils.R
\name{new_model_spec}
\alias{new_model_spec}
\title{Specification Utilities}
\usage{
new_model_spec(class, args, mode)
}
\description{
Internal constructors and engine helpers used to build model specifications
that mimic tidymodels-style objects.
}
\keyword{internal}
