% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_utils.R
\name{new_stochastic_vol_spec}
\alias{new_stochastic_vol_spec}
\title{Create a Stochastic Volatility Specification}
\usage{
new_stochastic_vol_spec(class, args, process_type, inheritance = character())
}
\description{
Create a Stochastic Volatility Specification
}
\keyword{internal}
