% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_models.R
\name{price_h1_hw_option_cos}
\alias{price_h1_hw_option_cos}
\title{Price H1-HW Options via COS Method}
\usage{
price_h1_hw_option_cos(
  process_spec,
  strikes,
  maturity,
  n_terms = 512L,
  truncation = 10,
  integration_points = 2000L
)
}
\arguments{
\item{process_spec}{A \code{bshw_spec} created by \code{\link[=bshw_spec]{bshw_spec()}}.}

\item{strikes}{Numeric vector of strikes.}

\item{maturity}{Numeric. Time to maturity.}

\item{n_terms}{Integer number of COS terms.}

\item{truncation}{Numeric truncation width.}

\item{integration_points}{Integer number of steps for the drift integral.}
}
\description{
Price H1-HW Options via COS Method
}
