% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_scholes.R
\name{price_options}
\alias{price_options}
\title{Price Options and Compute Greeks}
\usage{
price_options(model_spec, ...)
}
\arguments{
\item{model_spec}{Option specification object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A tibble containing model-dependent outputs.
}
\description{
Generic function for pricing option specifications and returning the full set
of analytic Greeks in tidy format.
}
