% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_generics.R
\name{price_zcb}
\alias{price_zcb}
\title{Price Fixed-Income Instruments}
\usage{
price_zcb(object, ...)
}
\arguments{
\item{object}{Model object.}

\item{...}{Forwarded to methods.}
}
\value{
Model-specific value.
}
\description{
Generic pricing dispatcher. Specific model objects provide concrete methods.
}
