% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_utils.R
\name{set_process_metadata}
\alias{set_process_metadata}
\title{Attach Metadata to a Process Specification}
\usage{
set_process_metadata(spec, ...)
}
\description{
Attach Metadata to a Process Specification
}
\keyword{internal}
