% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_utils.R
\name{set_spec_metadata}
\alias{set_spec_metadata}
\title{Attach Metadata to a Model Specification}
\usage{
set_spec_metadata(spec, ...)
}
\description{
Generic helper for specs that do not inherit from \code{process_spec} but still
need to publish engine information or auxiliary descriptors.
}
\keyword{internal}
