% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_paths_engines.R
\name{simulate_paths.abm_spec}
\alias{simulate_paths.abm_spec}
\title{Simulate ABM Paths (Engine Implementation)}
\usage{
\method{simulate_paths}{abm_spec}(process_spec, n_paths, n_steps, maturity, seed = 123, ...)
}
\description{
Simulate ABM Paths (Engine Implementation)
}
\keyword{internal}
