% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_paths_engines.R
\name{simulate_paths.gbm_spec}
\alias{simulate_paths.gbm_spec}
\title{Simulate GBM Paths (Engine Implementation)}
\usage{
\method{simulate_paths}{gbm_spec}(process_spec, n_paths, n_steps, maturity, seed = 123, ...)
}
\description{
Simulate GBM Paths (Engine Implementation)
}
\keyword{internal}
