% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_models.R
\name{szhw_characteristic_function}
\alias{szhw_characteristic_function}
\title{Schoebel-Zhu Hull-White Characteristic Function}
\usage{
szhw_characteristic_function(
  u,
  maturity,
  discount_fun,
  initial_rate,
  mean_reversion,
  short_rate_vol,
  initial_volatility,
  vol_mean_reversion,
  vol_long_term,
  vol_volatility,
  correlation_eq_vol,
  correlation_rate_vol,
  correlation_eq_rate,
  integration_points = 2000L
)
}
\arguments{
\item{u}{Numeric integration vector.}

\item{maturity}{Time to maturity.}

\item{discount_fun}{Discount function of the short-rate model.}

\item{initial_rate}{Initial short rate.}

\item{mean_reversion}{Hull-White mean reversion.}

\item{short_rate_vol}{Hull-White volatility (eta).}

\item{initial_volatility}{Initial level of the OU volatility factor.}

\item{vol_mean_reversion}{OU mean reversion.}

\item{vol_long_term}{OU long-run level.}

\item{vol_volatility}{OU volatility of volatility.}

\item{correlation_eq_vol}{Correlation between equity and volatility drivers.}

\item{correlation_rate_vol}{Correlation between short rate and volatility drivers.}

\item{correlation_eq_rate}{Correlation between equity and short rate drivers.}

\item{integration_points}{Trapezoid steps for deterministic integrals.}
}
\description{
Schoebel-Zhu Hull-White Characteristic Function
}
